- purge the hole cache when data is changed for now - fixed the unit tests because of the recent form token handling changes

+++ ACP3/Core/Cache/Purge.php
@@ -66,7 +66,7 @@ class Purge
 
             if (is_dir($path)) {
                 static::doPurge($path, $cacheId);
-                if (empty($cacheId)) {
+                if (empty($cacheId) && count(scandir($path)) == 2) {
                     rmdir($path);
                 }
             } elseif (empty($cacheId) || strpos($file, $cacheId) !== false) {
